created util query_set
