+++ app/utils/__init__.py
@@ -2,3 +2,4 @@ from .parser import parser_one_object, parser_all_object
 from .status_code import not_exist_code, exist_code
 from .pagination import paginate
 from .lists import contains
+from .query_set import RefQuerySet
